add input fields
