--- source/js/common/form/validationHelpers/IFocusableElement.js
@@ -1,16 +1,0 @@
-/**
- * Интерфейс для работы с фокусировкой объектов
- */
-export default class IFocusableElement {
-  init(el) {
-    throw 'IFocusableElement method @init not implemented';
-  }
-
-  setFocus() {
-    throw 'IFocusableElement method @setFocus not implemented';
-  }
-
-  inFocus() {
-		// optional method
-  }
-}
